Web/grapter/requirements.txt
fastapi==0.115.0
strawberry-graphql[fastapi]==0.243.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
pydantic<2.11